@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
-import 'rxjs/add/operator/map';
+import { map } from 'rxjs/operators';
 
 import { ApiHttp } from '../../shared/services/api-http.service';
 
@@ -11,11 +11,13 @@ export class ValidTokenGuard implements CanActivate {
               protected apiHttp: ApiHttp) { }
 
   canActivate() {
-    return this.apiHttp.currentTokenValid().map(valid => {
-      if(!valid) {
-        this.router.navigate(['/settings/']);
-      }
-      return valid;
-    });
+    return this.apiHttp.currentTokenValid().pipe(
+      map(valid => {
+        if(!valid) {
+          this.router.navigate(['/settings/']);
+        }
+        return valid;
+      })
+    );
   }
 }
