@@ -1,9 +1,8 @@
 import { Injectable } from '@angular/core';
-import { BehaviorSubject } from 'rxjs/BehaviorSubject';
-import 'rxjs/add/operator/map';
 
-import { City } from '../models/city.model'
-import { Observable } from 'rxjs/Observable';
+import { City } from '../models/city.model';
+import { Observable, BehaviorSubject } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 @Injectable()
 export class CurrentCity {
@@ -17,7 +16,9 @@ export class CurrentCity {
   }
 
   public currentCityValid(): Observable<boolean> {
-    return this.getCurrent().map(city => city !== undefined);
+    return this.getCurrent().pipe(
+      map(city => city !== undefined)
+    );
   }
 
   public setCity(city: City) {
