@@ -1,5 +1,7 @@
 import { Injectable } from '@angular/core';
 import { CanActivate, Router } from '@angular/router';
+import { Observable } from 'rxjs';
+import { map } from 'rxjs/operators';
 
 import { CurrentCity } from '../../shared/services/current-city.service';
 
@@ -9,12 +11,14 @@ export class ConfiguredCityGuard implements CanActivate {
   constructor(private router: Router,
               protected currentCity: CurrentCity) { }
 
-  canActivate() {
-    return this.currentCity.currentCityValid().map(valid => {
-      if(!valid) {
-        this.router.navigate(['/settings/']);
-      }
-      return valid;
-    });
+  canActivate(): Observable<boolean> {
+    return this.currentCity.currentCityValid().pipe(
+      map(valid => {
+        if(!valid) {
+          this.router.navigate(['/settings/']);
+        }
+        return valid;
+      })
+    );
   }
 }
